Externalize the application state management.
The application state (running, shutting down) is no more managed
by the JFrame internal state (visible, dispose....). So the frame
could be disposed to switch in fullscreen mode without shutting
down the artnet server.

For the time, the only registered listener is define in the main method.
This listener know/do to much things about/with the application components.
We should give the ApplicationState instance to all objects that need to do
something (like persisting data) when application state change.